+++ core/urls.py
@@ -13,14 +13,13 @@ urlpatterns = [
     # app
     path("api/", api.urls),
     path("project/<int:pk>/", views.ProjectDetailView.as_view(), name="project_detail"),
-    path("project/<int:pk>/settings/", views.ProjectSettingsView.as_view(), name="project_settings"),
-    path("blogging-agent/<int:pk>/", views.BloggingAgentDetailView.as_view(), name="blogging_agent_detail"),
-    path("pricing-agent/<int:pk>/", views.PricingAgentView.as_view(), name="pricing_agent"),
-    path("keywords-agent/<int:pk>/", views.KeywordsAgentView.as_view(), name="keywords_agent"),
     path(
-        "competitor-analysis-agent/<int:pk>/",
-        views.CompetitorAnalysisAgentView.as_view(),
-        name="competitor_analysis_agent",
+        "project/<int:pk>/settings/", views.ProjectSettingsView.as_view(), name="project_settings"
+    ),
+    path(
+        "project/<int:project_pk>/post/<int:pk>/",
+        views.GeneratedBlogPostDetailView.as_view(),
+        name="generated_blog_post_detail",
     ),
     # utils
     path("resend-confirmation/", views.resend_confirmation_email, name="resend_confirmation"),
@@ -31,6 +30,9 @@ urlpatterns = [
         views.create_checkout_session,
         name="user_upgrade_checkout_session",
     ),
-    path("create-customer-portal/", views.create_customer_portal_session, name="create_customer_portal_session"),
-    # Add these with your other URLs
+    path(
+        "create-customer-portal/",
+        views.create_customer_portal_session,
+        name="create_customer_portal_session",
+    ),
 ]