+++ CHANGELOG.md
@@ -13,6 +13,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.2] - 2025-08-10
+**Removed**
+- Various agents, focus on generating blog post content
+- PricingPageAnalysis and CompetitorBlog post models
+- Views for now unsporrted models
+
+**Changed**
+- Simplification of the design and the UI
+
+**Added**
+- A page for generated content
+
 ## [0.0.1] - 2025-08-09
 **Added**
 - Adding automated posting feature